backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
azure-identity>=1.15.0
azure-search-documents>=11.4.0
azure-ai-formrecognizer>=3.3.0
openai>=1.12.0
pydantic>=2.5.3,<3.0.0
python-multipart>=0.0.6

